refactor to move all tfa logic into separate class
